pyproject.toml
[project]
name = "tts-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chatterbox-tts>=0.1.1",
    "setuptools>=80.9.0",
]
